Make function declarations & definitions agree

+++ src/usolve/libusolve.h
@@ -32,8 +32,8 @@ unsigned long int mpz_poly_max_bsize_coeffs(mpz_t *, unsigned long int);
 unsigned long int mpz_poly_min_bsize_coeffs(mpz_t *, unsigned long int);
 
 void mpz_poly_eval_2exp_naive(mpz_t *,
-                              unsigned long int,
-                              mpz_t *, const int,
+                              long int,
+                              mpz_t *, const long int,
                               mpz_t *, mpz_t *);
 
 void mpz_poly_eval_2exp_naive2(mpz_t *,
